refactor styles to css module
