+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/PipetteFields/TipSelectionWizard/PipetteShadows/PipetteLabel/index.ts
@@ -1,0 +1,1 @@
+export { PipetteLabel } from './PipetteLabel'
